salt: add procps for better process controll support in modules
procps is a set of command line and full-screen utilities that
provide information out of the pseudo-filesystem most commonly
located at /proc. This filesystem provides a simple interface
to the kernel data structures.
The programs of procps generally concentrate on the structures
that describe the processess running on the system.